[ENH] implement Dave's review comments
