+++ api/app/importer/mapping.py
@@ -1,8 +1,14 @@
+import csv
 import json
+import os
 from os.path import isfile
-import re
+from typing import List, Dict
+from core.models import Plate, PlateMapping
+from core.mapping import Mapping, MappingList
+from django.core.files.base import ContentFile
 
 import yaml
+from friendlylog import colored_logger as log
 
 from .helper import sameSchema
 
@@ -88,20 +94,95 @@ class EchoMapping():
         "current_fluid_volume": "Current Fluid Volume",
         "percent_dms": "% DMSO"}
 
+
+
     @staticmethod
+    def read_csv_echo_file(reader: csv.reader, headers) -> List[
+        Dict[str, str]]:
+        """
+        Reads a CSV file and returns a list of dictionaries representing each row in the file.
+        """
+        table_data = []
+        column_headers = list(headers.values())
+        for row_index, table_row in enumerate(reader):
+            if len(table_row) > 0 and table_row[0] == 'Source Plate Name':
+                column_headers = table_row
 
+            if row_index > 9:
+                row_object = {}
+                for cell_index, cell in enumerate(table_row):
+                    row_object[column_headers[cell_index]] = cell
+                table_data.append(row_object)
+        return table_data
 
+    @staticmethod
+    def get_csv_echo_files(path: str, headers) -> list[
+        dict[str, list[dict[str, str]] | str]]:
+        """
+        Recursively searches the given directory for CSV files with "transfer" in their name,
+        reads each file using the read_csv_file function, and returns a list of tables
+        representing the data in the CSV files.
+        """
+        files_data = []
+        for root, dirs, files in os.walk(path, topdown=False):
+            for file in files:
+                if file.endswith(".csv") and 'transfer' in file:
+                    subdirectory = os.path.relpath(root, path)
+                    file_path = os.path.join(path, subdirectory, file)
+                    with open(os.path.join(root, file), 'r') as csv_file:
+                        log.debug(f"Reading file {file}")
+                        csv_reader = csv.reader(csv_file, delimiter=',')
+                        files_data.append({"file_path": file_path,
+                                              "data": EchoMapping.read_csv_echo_file(
+                                                  csv_reader, headers)})
+        return files_data
 
-    def convert_position_to_index(position: str,
-                                number_of_columns: int) -> int:
-        match = re.match(r'([a-zA-Z]+)(\d+)', position)
-        letters = match.group(1)
-        col = int(match.group(2))
-        row = 0
-        for char in letters:
-            row += ord(char.upper()) - 65
+    @staticmethod
+    def parse_plate_data(mapping_data, file_path, headers, experiment_name):
+        source_plate_barcode = mapping_data[0][headers['source_plate_barcode']]
+        destination_plate_barcode = mapping_data[0][
+            headers['destination_plate_barcode']]
 
-        index = row * number_of_columns + + (col - 1)
-        return index
+        source_plate = Plate.objects.get(barcode=source_plate_barcode)
+        destination_plate = Plate.objects.get(
+            barcode=destination_plate_barcode)
 
+        if not destination_plate:
+            raise ValueError(
+                f"Destination plate with barcode '{destination_plate_barcode}' "
+                f"not found. Please add plates to the experiment.")
+        number_of_columns = destination_plate.dimension.cols
 
+        if destination_plate.experiment.name != experiment_name:
+            raise ValueError(
+                f"Destination plate with barcode '{destination_plate_barcode}' "
+                f"does not belong to experiment '{experiment_name}'.")
+
+        mapping_list = MappingList()
+        for entry in mapping_data:
+            if len(entry) > 0:
+                log.debug(f"Mapping {entry[headers['source_well']]} to "
+                          f"{entry[headers['destination_well']]} from"
+                          f" {source_plate_barcode} to {destination_plate_barcode}")
+
+                from_pos = Plate.convert_position_to_index(
+                    entry[headers['source_well']], number_of_columns)
+                to_pos = Plate.convert_position_to_index(
+                    entry[headers['destination_well']], number_of_columns)
+                mapping = Mapping(from_pos=from_pos, to_pos=to_pos,
+                                  amount=float(
+                                      entry[headers['actual_volume']]))
+                mapping_list.add(mapping)
+        mapping_success = source_plate.map(mapping_list, destination_plate)
+        if mapping_success:
+            file_content = open(file_path, 'rb').read()
+            file_name = os.path.basename(file_path)
+            plate_mapping = PlateMapping(source_plate=source_plate,
+                                         target_plate=destination_plate)
+            plate_mapping.mapping_file.save(file_name,
+                                            ContentFile(file_content))
+            log.info(f"Successfully mapped {source_plate_barcode} to "
+                     f"{destination_plate_barcode}")
+        else:
+            log.error(f"Failed to map {source_plate_barcode} to "
+                      f"{destination_plate_barcode}")
